The partitioned index will now perform repeated compacting merges on a named index.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@129 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/test/com/bigdata/journal/AbstractTestCase.java
@@ -483,13 +483,13 @@ abstract public class AbstractTestCase
             File tmp = File.createTempFile("test-" + bufferMode + "-"
                     + name + "-", ".jnl");
             
-//            // Delete the file otherwise the Journal will attempt to open it.
-//            if (!tmp.delete()) {
-//
-//                throw new RuntimeException("Unable to remove empty test file: "
-//                        + tmp);
-//
-//            }
+            // Delete the file otherwise the Journal will attempt to open it.
+            if (!tmp.delete()) {
+
+                throw new RuntimeException("Unable to remove empty test file: "
+                        + tmp);
+
+            }
 
             // make sure that the file is eventually removed.
             tmp.deleteOnExit();
